% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palec.R
\name{Octaves}
\alias{Octaves}
\title{Assign species to octave}
\usage{
Octaves(counts)
}
\arguments{
\item{counts}{Vector containing number of individuals in each species.
Zero entries will be ignored.}
}
\value{
Vector listing which octave each non-zero species belongs to,
suitable for the constrution of a histogram (with \code{\link[=hist]{hist()}}) or
tabulation (with \code{\link[=table]{table()}}).
}
\description{
For details of octave plots, see \href{https://smithlabdurham.github.io/GEOL2031/2A.html#Octave}{course material}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
