% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palec.R
\name{palec}
\alias{palec}
\alias{Survivorship}
\alias{Histogram}
\alias{Diversity}
\title{Palaeoecological apps}
\usage{
Survivorship()

Histogram()

Diversity()
}
\description{
Applications for palaeontological analysis
}
\details{
These applications accompany the GEOL2031 course; an introduction to their
use is available in the \href{https://smithlabdurham.github.io/GEOL2031/2S.html}{course documents}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
