% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palec.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{palettes}
\format{
An object of class \code{list} of length 21.
}
\usage{
palettes
}
\description{
A list of colourblind-readable palettes generated using
\href{https://medialab.github.io/iwanthue/}{iwanthue}.
}
\details{
Entry \code{i} in the list contains \code{i} discriminable colours in six-digit
hexadecimal RGB format.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{datasets}
